// ui/src/api/generated/models/test-rag-source.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.26.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TestRagSource
 */
export interface TestRagSource {
    /**
     * 
     * @type {string}
     * @memberof TestRagSource
     */
    'content'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestRagSource
     */
    'id'?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof TestRagSource
     */
    'metadata'?: { [key: string]: object; };
    /**
     * 
     * @type {number}
     * @memberof TestRagSource
     */
    'score'?: number;
    /**
     * 
     * @type {string}
     * @memberof TestRagSource
     */
    'sourceType'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestRagSource
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestRagSource
     */
    'url'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TestRagSource
     */
    'usedForContext'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TestRagSource
     */
    'visible'?: boolean;
}